requirements.txt
boto3>=1.34.0
botocore>=1.34.0
cdflib>=1.2.0
geopack>=1.0.10
keyring>=24.2.0
matplotlib>=3.8.0
numba>=0.59.0
numpy>=1.22
pandas>=2.2.0
pycdfpp>=0.6.0
python-dateutil>=2.9.0
requests>=2.31.0
scipy>=1.11.2
tqdm>=4.66.2
xarray>=2024.1.0
